backend/requirements.txt
# Wani Backend - Python Dependencies
# Python 3.13+ compatible
# All versions are pinned for consistency across environments

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
email-validator==2.3.0  # Required for Pydantic EmailStr

# Database
sqlalchemy==2.0.44
psycopg[binary,pool]==3.2.10  # Reemplaza asyncpg - compatible con Python 3.13+
alembic==1.17.0
supabase==2.10.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.12
slowapi==0.1.9  # Rate limiting for FastAPI

# Redis & Caching
redis==5.0.1
hiredis==3.3.0

# Celery (Background Jobs)
celery==5.3.4

# Blockchain (Stellar)
stellar-sdk==9.1.0

# External Integrations
requests==2.31.0
twilio==8.10.0
sendgrid==6.10.0
httpx==0.27.2
resend==2.6.0  # Modern email service for transactional emails

# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0

# Data Processing
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Type Stubs
types-redis==4.6.0.11
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
